+++ packages/core/src/icons/components/index.js
@@ -1,8 +1,12 @@
 import AddItem from './AddItem';
+import AddLocation from './AddLocation';
 import AddMultiItems from './AddMultiItems';
 import Add from './Add';
 import Announce from './Announce';
 import Cash from './Cash';
+import CheckboxEmpty from './CheckboxEmpty';
+import CheckboxHalf from './CheckboxHalf';
+import CheckboxSelected from './CheckboxSelected';
 import ClearItem from './ClearItem';
 import CopyLink from './CopyLink';
 import Copypaste from './Copypaste';
@@ -43,6 +47,7 @@ import InventoryCategory from './InventoryCategory';
 import InventoryItem from './InventoryItem';
 import LastPage from './LastPage';
 import Loading from './Loading';
+import Location from './Location';
 import Lock from './Lock';
 import Mapping from './Mapping';
 import MinusOne from './MinusOne';
@@ -61,7 +66,6 @@ import Prev from './Prev';
 import Printer from './Printer';
 import Question from './Question';
 import RadioEmpty from './RadioEmpty';
-import RadioHalf from './RadioHalf';
 import RadioSelected from './RadioSelected';
 import Refresh from './Refresh';
 import RemoveElement from './RemoveElement';
@@ -78,10 +82,14 @@ import Unlock from './Unlock';
 
 export default {
   'add-item': AddItem,
+  'add-location': AddLocation,
   'add-multi-items': AddMultiItems,
   add: Add,
   announce: Announce,
   cash: Cash,
+  'checkbox-empty': CheckboxEmpty,
+  'checkbox-half': CheckboxHalf,
+  'checkbox-selected': CheckboxSelected,
   'clear-item': ClearItem,
   'copy-link': CopyLink,
   copypaste: Copypaste,
@@ -122,6 +130,7 @@ export default {
   'inventory-item': InventoryItem,
   'last-page': LastPage,
   loading: Loading,
+  location: Location,
   lock: Lock,
   mapping: Mapping,
   'minus-one': MinusOne,
@@ -140,7 +149,6 @@ export default {
   printer: Printer,
   question: Question,
   'radio-empty': RadioEmpty,
-  'radio-half': RadioHalf,
   'radio-selected': RadioSelected,
   refresh: Refresh,
   'remove-element': RemoveElement,