+++ packages/form/src/index.js
@@ -4,6 +4,8 @@ import SelectRow from './SelectRow';
 
 import SelectList from './SelectList';
 import SelectOption from './SelectOption';
+import RadioSelectList from './RadioSelectList';
+import RadioSelectOption from './RadioSelectOption';
 import parseSelectOptions from './utils/parseSelectOptions';
 
 export {
@@ -14,4 +16,6 @@ export {
   SelectList,
   SelectOption,
   parseSelectOptions,
+  RadioSelectList,
+  RadioSelectOption,
 };