+++ CHANGELOG.md
@@ -6,6 +6,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Changed
+- [Core] Should not apply asideControlClickableOnDisabled if not disabled. (#332)
+
+### Added
+- [Core] Add new icon: Location. (#333)
+- [Core] Add Radio component (#338, #339)
+- [Form] Add RadioSelectList and RadioSelectOption (#340)
+
+### Breaking
+- [Core] Rename radio-empty.svg to checkbox-empty.svg (#337)
+- [Core] Rename radio-half.svg to checkbox-half.svg (#337)
+- [Core] Rename radio-selected.svg to checkbox-selected.svg (#337)
+- [Core] Fix the svgr/cli version at v5.5.0(#347)
+
 ## [4.4.0]
 
 ### Breaking